+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 - Bump JavaScript SDK from v8.53.0 to v8.54.0 ([#4503](https://github.com/getsentry/sentry-react-native/pull/4503))
   - [changelog](https://github.com/getsentry/sentry-javascript/blob/develop/CHANGELOG.md#8540)
   - [diff](https://github.com/getsentry/sentry-javascript/compare/8.53.0...8.54.0)
+- Bump Bundler Plugins from v2.20.1 to v3.1.2 ([#4519](https://github.com/getsentry/sentry-react-native/pull/4519))
+  - [changelog](https://github.com/getsentry/sentry-javascript-bundler-plugins/blob/main/CHANGELOG.md#312)
+  - [diff](https://github.com/getsentry/sentry-javascript-bundler-plugins/compare/2.20.1...3.1.2)
 
 ## 6.6.0